backend/pyproject.toml
[project]
name = "hackathon-final"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "dotenv>=0.9.9",
    "fastapi>=0.115.0",
    "fpdf>=1.7.2",
    "langchain>=1.2.3",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.1",
    "langchain-ollama>=1.0.1",
    "langchain-pinecone>=0.2.13",
    "langchain-text-splitters>=1.1.0",
    "pinecone>=7.3.0",
    "pytube>=15.0.0",
    "uvicorn>=0.32.0",
    "youtube-transcript-api>=1.2.3",
    "pymongo[srv]>=4.6.0",
    "google>=3.0.0",
    "grpcio>=1.76.0",
    "protobuf>=6.33.4",
    "grpcio-tools>=1.76.0",
    "google-api-python-client>=2.188.0",
    "motor>=3.7.1",
    "pydantic[email]==2.12.5",
    "httpx>=0.28.1",
    "google-cloud-aiplatform>=1.134.0",
    "cloudpickle==3.1.2",
    "google-auth>=2.29.0",
    "python-jose[cryptography]>=3.3.0",
]

[tool.setuptools.packages.find]
include = ["models*", "services*"]
